created and integrated RabbitMQConnector class for connecting with RabbitMQ and moved tests to ./tests
