+++ package.json
@@ -5,7 +5,8 @@
   "type": "module",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "node ./tests/test.js",
+    "start": "node index.js"
   },
   "repository": {
     "type": "git",
@@ -19,6 +20,7 @@
   },
   "homepage": "https://github.com/YAA-C/Uploader#readme",
   "dependencies": {
+    "amqplib": "^0.10.3",
     "dotenv": "^16.4.5",
     "mongoose": "^8.2.1"
   }
